+++ internal/e2e-client/tests/callfabric/videoRoom.spec.ts
@@ -1,8 +1,5 @@
 import { uuid } from '@signalwire/core'
-import {
-  UnifiedCommunicationSession,
-  CallJoinedEventParams,
-} from '@signalwire/client'
+import { CallSession, CallJoinedEventParams } from '@signalwire/client'
 import { test, expect } from '../../fixtures'
 import {
   SERVER_URL,
@@ -47,7 +44,7 @@ test.describe('CallFabric VideoRoom', () => {
     await page.evaluate(
       async ({ roomSession }) => {
         // @ts-expect-error
-        const roomObj: UnifiedCommunicationSession = window._roomObj
+        const roomObj: CallSession = window._roomObj
 
         const memberUpdatedMuted = new Promise((resolve) => {
           const memberUpdatedEvent = new Promise((res) => {
@@ -117,7 +114,7 @@ test.describe('CallFabric VideoRoom', () => {
     await page.evaluate(
       async ({ roomSession }) => {
         // @ts-expect-error
-        const roomObj: UnifiedCommunicationSession = window._roomObj
+        const roomObj: CallSession = window._roomObj
 
         const memberUpdatedMuted = new Promise((resolve) => {
           const memberUpdatedEvent = new Promise((res) => {
@@ -192,7 +189,7 @@ test.describe('CallFabric VideoRoom', () => {
     // --------------- Screenshare ---------------
     await page.evaluate(async () => {
       // @ts-expect-error
-      const roomObj: UnifiedCommunicationSession = window._roomObj
+      const roomObj: CallSession = window._roomObj
 
       let screenMemberId: string | undefined
 
@@ -246,7 +243,7 @@ test.describe('CallFabric VideoRoom', () => {
       // @ts-expect-error
       async ({ roomSession }) => {
         // @ts-expect-error
-        const roomObj: UnifiedCommunicationSession = window._roomObj
+        const roomObj: CallSession = window._roomObj
 
         const roomUpdatedLocked = new Promise((resolve) => {
           roomObj.on('room.updated', (params) => {