+++ packages/client/src/VideoOverlays.ts
@@ -1,12 +1,9 @@
 import {
-  FabricMemberUpdatedEventParams as UnifiedCommunicationMemberUpdatedEventParams,
-  getLogger
+  FabricMemberUpdatedEventParams as CallMemberUpdatedEventParams,
+  getLogger,
 } from '@signalwire/core'
 import { VideoRoomSession, isVideoRoomSession } from './video/VideoRoomSession'
-import {
-  UnifiedCommunicationSession,
-  isUnifiedCommunicationSession,
-} from './fabric/UnifiedCommunicationSession'
+import { CallSession, isCallSession } from './fabric/CallSession'
 import { VideoMemberUpdatedHandlerParams } from './utils/interfaces'
 import { OVERLAY_PREFIX, SDK_PREFIX } from './utils/roomSession'
 
@@ -69,12 +66,12 @@ export class UserOverlay {
 interface LocalVideoOverlayOptions {
   id: string
   mirrorLocalVideoOverlay: boolean
-  room: UnifiedCommunicationSession | VideoRoomSession
+  room: CallSession | VideoRoomSession
 }
 
 export class LocalVideoOverlay extends UserOverlay {
   private _mirrored: boolean
-  private _room: UnifiedCommunicationSession | VideoRoomSession
+  private _room: CallSession | VideoRoomSession
 
   constructor(options: LocalVideoOverlayOptions) {
     super(options)
@@ -99,7 +96,7 @@ export class LocalVideoOverlay extends UserOverlay {
   }
 
   private attachListeners() {
-    if (isUnifiedCommunicationSession(this._room)) {
+    if (isCallSession(this._room)) {
       this._room.on(
         'member.updated.videoMuted',
         this.fabricMemberVideoMutedHandler
@@ -114,7 +111,7 @@ export class LocalVideoOverlay extends UserOverlay {
 
   /** @internal */
   public detachListeners() {
-    if (isUnifiedCommunicationSession(this._room)) {
+    if (isCallSession(this._room)) {
       this._room.off(
         'member.updated.videoMuted',
         this.fabricMemberVideoMutedHandler
@@ -137,9 +134,7 @@ export class LocalVideoOverlay extends UserOverlay {
     }
   }
 
-  private fabricMemberVideoMutedHandler(
-    params: UnifiedCommunicationMemberUpdatedEventParams
-  ) {
+  private fabricMemberVideoMutedHandler(params: CallMemberUpdatedEventParams) {
     this.memberVideoMutedHandler(
       params.member.member_id,
       params.member.video_muted