+++ packages/client/src/fabric/interfaces/incomingCallManager.ts
@@ -1,4 +1,4 @@
-import { UnifiedCommunicationSession } from '../UnifiedCommunicationSession'
+import { CallSession } from '../CallSession'
 import { CallParams } from './wsClient'
 
 export type IncomingInviteSource = 'websocket' | 'pushNotification'
@@ -21,7 +21,7 @@ export interface IncomingInviteWithSource extends IncomingInvite {
 export interface IncomingCallNotification {
   invite: {
     details: IncomingInvite
-    accept: (param: CallParams) => Promise<UnifiedCommunicationSession>
+    accept: (param: CallParams) => Promise<CallSession>
     reject: () => Promise<void>
   }
 }